refactor(daily): move `/multi-images` page and components
